# docs/api/swagger.yaml
definitions:
  APIErrorResponse:
    properties:
      code:
        type: integer
      details: {}
      message:
        type: string
      status:
        type: string
    type: object
...
  github_com_flotio-dev_core-api_internal_common_database.ProjectConfig:
    properties:
      android_build_args:
        type: string
      android_build_format:
        description: '"apk" or "aab"'
        type: string
      build_mode:
        description: '"debug", "release", "profile"'
        type: string
      build_trigger:
        type: string
      cocoapods_version:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      dependency_caching:
        type: boolean
      dependency_dirs:
        items:
          type: string
        type: array
      do_not_send_for_review:
        type: boolean
      email_recipients:
        items:
          type: string
        type: array
      enable_android_code_signing:
        type: boolean
      enable_email_notifications:
        description: Notifications
        type: boolean
      enable_flutter_analyze:
        type: boolean
      enable_flutter_driver:
        type: boolean
      enable_flutter_test:
        type: boolean
      enable_google_play_publishing:
        type: boolean
      flutter_analyze_args:
        type: string
      flutter_driver_args:
        type: string
      flutter_driver_targets:
        items:
          type: string
        type: array
      flutter_test_args:
        type: string
      flutter_version:
        description: Build
        type: string
      git_repo:
        description: Git Connection
        type: string
      git_token:
        type: string
      git_username:
        type: string
      google_play_credentials_id:
        type: integer
      google_play_track:
        type: string
      id:
        type: integer
      ios_build_args:
        type: string
      keystore_id:
        description: Linked User-level Assets
        type: integer
      package_name:
        description: Distribution
        type: string
      platforms:
        items:
          type: string
        type: array
      post_build_script:
        type: string
      post_clone_script:
        description: Scripts
        type: string
      post_test_script:
        type: string
      pre_build_script:
        type: string
      pre_publish_script:
        type: string
      pre_test_script:
        type: string
      project_id:
        type: integer
      project_path:
        type: string
      publish_even_if_tests_fail:
        type: boolean
      rollout_fraction:
        type: number
      submit_as_draft:
        type: boolean
      test:
        description: Testing
        type: boolean
      update_priority:
        type: integer
      updatedAt:
        type: string
      watched_branch_patterns:
        items:
          $ref: '#/definitions/github_com_flotio-dev_core-api_internal_common_database.WatchedBranchPattern'
        type: array
      watched_tag_patterns:
        items:
          $ref: '#/definitions/github_com_flotio-dev_core-api_internal_common_database.WatchedTagPattern'
        type: array
      web_build_args:
        type: string
      webhook_urls:
        description: Webhooks
        items:
          type: string
        type: array
      xcode_version:
        type: string
    type: object
  github_com_flotio-dev_core-api_internal_common_database.WatchedBranchPattern:
    properties:
      pattern:
        type: string
      target:
        description: '"source" or "target"'
        type: string
      type:
        description: '"include" or "exclude"'
        type: string
    type: object
  github_com_flotio-dev_core-api_internal_common_database.WatchedTagPattern:
    properties:
      pattern:
        type: string
      type:
        description: '"include" or "exclude"'
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_modules_project_handler.DeleteResponse:
    properties:
      status:
        type: string
    type: object
...
  internal_modules_project_handler.ProjectConfigResponse:
    properties:
      config:
        $ref: '#/definitions/github_com_flotio-dev_core-api_internal_common_database.ProjectConfig'
    type: object
...
info:
  contact:
    email: support@flotio.dev
    name: Flotio Team
  description: Flotio Core API — authentication, environment/keystore/Google Play
    credentials assets, projects, build pipeline, releases and GitHub integration
    for the Flotio mobile CI platform.
  license:
    name: Proprietary
  title: Flotio Core API
  version: 1.0.0
paths:
...
  /project/{id}/config:
    delete:
      consumes:
      - application/json
      description: Delete or reset the configuration for a specific project
      operationId: ConfigDeleteHandler
      parameters:
      - description: Project ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_project_handler.DeleteResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete project configuration
      tags:
      - config
    get:
      consumes:
      - application/json
      description: Get the configuration for a specific project
      operationId: ConfigGetHandler
      parameters:
      - description: Project ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_project_handler.ProjectConfigResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get project configuration
      tags:
      - config
    post:
      consumes:
      - application/json
      description: Create or update the configuration for a specific project (Supports
        partial updates)
      operationId: ConfigPostHandler
      parameters:
      - description: Project ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: Configuration data
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/github_com_flotio-dev_core-api_internal_common_database.ProjectConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_project_handler.ProjectConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Update project configuration
      tags:
      - config
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"